Move generator logic to it's own class Thi should make testing easier
